refactor: move main page components
