+++ src/components/main/Showplant.js
